requirements.txt
async-generator==1.10
attrs==22.1.0
blinker==1.5
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
cryptography==37.0.4
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
iso8601==1.0.2
kaitaistruct==0.10
m3u8==3.0.0
outcome==1.2.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0
pydivert==2.1.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
requests==2.28.1
selenium==4.3.0
selenium-wire==4.6.5
sniffio==1.2.0
sortedcontainers==2.4.0
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
wsproto==1.1.0
zstandard==0.18.0
